backend/requirements.txt
Flask==3.1.0
Flask-JWT-Extended==4.7.1
Flask-CORS==5.0.1
PyMySQL==1.1.1
Werkzeug==3.1.3
python-dotenv==1.1.0
